server/requirements.txt
dnspython==2.1.0
Flask==2.0.1
python-dotenv==0.19.0
pymongo==3.12.0
gunicorn==20.0.4
flask-cors==3.0.10